requirements.txt
# Core dependencies
numpy
langchain
langchain-community
langchain-core
llama-cpp-python

# Graph and workflow dependencies
langgraph
pydantic
typing-extensions

# CLI and interface dependencies
rich
prompt-toolkit

# Utility dependencies
python-dotenv
pyyaml